requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-groq>=0.2.0
langgraph>=0.2.0
langsmith>=0.1.0
python-dotenv==1.0.0
requests==2.31.0
pydantic>=2.8.0
streamlit>=1.38.0
plotly>=5.18.0
